[step4] fix CarRacingController 객체 생성자 수정 및 이에 따른 테스트 코드 수정
